User review and rating system

+++ inventory/models.py
@@ -54,7 +54,7 @@ class UserProfile(models.Model):
 class Review(models.Model):
     car = models.ForeignKey(Car, on_delete=models.CASCADE, related_name='reviews')
     user = models.ForeignKey(User, on_delete=models.CASCADE)
-    rating = models.PositiveIntegerField(choices=[(i, i) for i in range(1, 6)]) # ratings between 1 and 5
+    rating = models.PositiveSmallIntegerField(choices=[(i, i) for i in range(1, 6)]) # ratings between 1 and 5
     comment = models.TextField(blank=True)
     created_at = models.DateTimeField(auto_now_add=True)
